requirements.txt

mcp
asyncpg
python-dotenv
pydantic
fastapi
uvicorn
psycopg2-binary
